refactor: use named exports for UI components
