+++ app/components/Activity/InvoiceModal/InvoiceModal.js
@@ -4,7 +4,7 @@ import QRCode from 'qrcode.react'
 import copy from 'copy-to-clipboard'
 import { showNotification } from 'lib/utils/notifications'
 import Value from 'components/Value'
-import Dropdown from 'components/UI/Dropdown'
+import { Dropdown } from 'components/UI'
 import { FormattedDate, FormattedTime, FormattedMessage } from 'react-intl'
 import Countdown from '../Countdown'
 import messages from './messages'
