+++ app/components/Contacts/SubmitChannelForm/SubmitChannelForm.js
@@ -1,15 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-
 import FaExclamationCircle from 'react-icons/lib/fa/exclamation-circle'
-
 import AmountInput from 'components/AmountInput'
-import Button from 'components/UI/Button'
-import Dropdown from 'components/UI/Dropdown'
-
+import { Button, Dropdown } from 'components/UI'
 import { FormattedNumber, FormattedMessage } from 'react-intl'
 import messages from './messages'
-
 import styles from './SubmitChannelForm.scss'
 
 class SubmitChannelForm extends React.Component {
