+++ app/components/Form/Request/Request.js
@@ -1,16 +1,11 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-
-import Hand from 'components/Icon/Hand'
-
 import { btc } from 'lib/utils'
+import Hand from 'components/Icon/Hand'
 import AmountInput from 'components/AmountInput'
-import Button from 'components/UI/Button'
-import Dropdown from 'components/UI/Dropdown'
-
+import { Button, Dropdown } from 'components/UI'
 import { FormattedNumber, FormattedMessage, injectIntl } from 'react-intl'
 import messages from './messages'
-
 import styles from './Request.scss'
 
 const Request = ({
