+++ app/components/Wallet/Wallet.js
@@ -6,10 +6,7 @@ import { Box, Flex } from 'rebass'
 import { btc, blockExplorer } from 'lib/utils'
 import Value from 'components/Value'
 import Settings from 'components/Settings'
-import Button from 'components/UI/Button'
-import Text from 'components/UI/Text'
-import Dropdown from 'components/UI/Dropdown'
-
+import { Button, Dropdown, Text } from 'components/UI'
 import CheckAnimated from 'components/Icon/CheckAnimated'
 import ZapLogo from 'components/Icon/ZapLogo'
 import ZapLogoBlack from 'components/Icon/ZapLogoBlack'
