+++ stories/components/background.stories.js
@@ -1,8 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import BackgroundDark from 'components/UI/BackgroundDark'
-import BackgroundLight from 'components/UI/BackgroundLight'
-import BackgroundLightest from 'components/UI/BackgroundLightest'
+import { BackgroundDark, BackgroundLight, BackgroundLightest } from 'components/UI'
 
 storiesOf('Components.Background', module)
   .add('dark', () => (
