+++ stories/components/bar.stories.js
@@ -1,5 +1,5 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
-import Bar from 'components/UI/Bar'
+import { Bar } from 'components/UI'
 
 storiesOf('Components.Bar', module).add('bar', () => <Bar />)
