+++ stories/components/spinner.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
 import { storiesOf } from '@storybook/react'
-import Spinner from 'components/UI/Spinner'
+import { Spinner } from 'components/UI'
 
 storiesOf('Components.Spinners', module).add('circle', () => <Spinner />)
