+++ test/unit/components/UI/BackgroundDark.spec.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import BackgroundDark from 'components/UI/BackgroundDark'
 import renderer from 'react-test-renderer'
+import { BackgroundDark } from 'components/UI'
 
 describe('component.UI.BackgroundDark', () => {
   it('should render correctly', () => {
