+++ test/unit/components/UI/BackgroundLightest.spec.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import BackgroundLightest from 'components/UI/BackgroundLightest'
 import renderer from 'react-test-renderer'
+import { BackgroundLightest } from 'components/UI'
 
 describe('component.UI.BackgroundLightest', () => {
   it('should render correctly', () => {
