+++ test/unit/components/UI/Button.spec.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Button from 'components/UI/Button'
 import renderer from 'react-test-renderer'
+import { Button } from 'components/UI'
 
 describe('component.UI.Button', () => {
   it('should render correctly', () => {
