+++ test/unit/components/UI/Modal.spec.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Modal from 'components/UI/Modal'
 import renderer from 'react-test-renderer'
+import { Modal } from 'components/UI'
 
 describe('component.UI.Modal', () => {
   it('should render correctly', () => {
