+++ test/unit/components/UI/Page.spec.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Page from 'components/UI/Page'
 import renderer from 'react-test-renderer'
+import { Page } from 'components/UI'
 
 describe('component.UI.Page', () => {
   it('should render correctly', () => {
