+++ test/unit/components/UI/Toggle.spec.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Form } from 'informed'
-import Toggle from 'components/UI/Toggle'
 import renderer from 'react-test-renderer'
 import { dark } from 'themes'
 import { ThemeProvider } from 'styled-components'
+import { Toggle } from 'components/UI'
 
 describe('component.UI.Toggle', () => {
   it('should render correctly', () => {
